Connected home screen to login/create account buttons. - created JFrame for home screen
