--- homeScreen.form
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="homeScreen">
-  <grid id="27dc6" binding="homePanel" layout-manager="GridLayoutManager" row-count="3" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="5" y="20" width="846" height="516"/>
-    </constraints>
-    <properties>
-      <background color="-16777216"/>
-    </properties>
-    <border type="none"/>
-    <children>
-      <component id="dc7ff" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="9" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <foreground color="-16718561"/>
-          <text value="MnM Uni Planner"/>
-        </properties>
-      </component>
-      <component id="762da" class="javax.swing.JButton" binding="xButton" default-binding="true">
-        <constraints>
-          <grid row="0" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="5" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <background color="-3881780"/>
-          <text value="X"/>
-        </properties>
-      </component>
-      <component id="bd5ac" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="2" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="6" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <foreground color="-16718561"/>
-          <text value="MnM Uni Planner"/>
-        </properties>
-      </component>
-      <component id="88241" class="javax.swing.JLabel">
-        <constraints>
-          <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <font name="Courier New" size="48" style="1"/>
-          <foreground color="-16718561"/>
-          <text value="Hello, Name."/>
-        </properties>
-      </component>
-      <component id="8e2ae" class="javax.swing.JButton" binding="classes">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <background color="-1576705"/>
-          <font name="Courier New" size="26" style="1"/>
-          <text value="Classes"/>
-        </properties>
-      </component>
-      <component id="9354a" class="javax.swing.JButton" binding="calendar">
-        <constraints>
-          <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <background color="-1707777"/>
-          <font name="Courier New" size="26" style="1"/>
-          <text value="Calendar"/>
-        </properties>
-      </component>
-      <component id="1a060" class="javax.swing.JButton" binding="studyGuide">
-        <constraints>
-          <grid row="1" column="2" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <background color="-1707777"/>
-          <font name="Courier New" size="26" style="1"/>
-          <text value="StudyGuide"/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
