+++ src/OpeningScreen.java
@@ -103,6 +103,8 @@ public class OpeningScreen extends JFrame implements ActionListener{
         else if(e.getSource() == loginToMenuButton){
             String username = usernameField.getText();
             String password = passwordField.getText();
+            new homeScreen();
+            display.dispose();
             // proceed to home screen after checking existing files
         }
         else if(e.getSource() == createAccountButton){
@@ -112,7 +114,8 @@ public class OpeningScreen extends JFrame implements ActionListener{
             String confirmPassword = confirmPasswordField.getText();
             boolean valid = checkRegisterInfo(name, username, password, confirmPassword);
             if(valid){
-                // proceed to home screen
+                new homeScreen();
+                display.dispose();
             }
         }
     }
